Fix StreamingService/launch
- Interpret code 130 (128+2) as a success code
- Rewrite tests and use sinon

+++ src/app/services/StreamingService/launch/index.js
@@ -74,7 +74,7 @@ export default async function( stream, provider, player, onSuccess ) {
 					return launch();
 				}
 
-				if ( code === 0 ) {
+				if ( code === 0 || code === 130 ) {
 					set( stream, "isCompleted", true );
 					resolve();
 				} else if ( signal ) {
@@ -128,7 +128,7 @@ export default async function( stream, provider, player, onSuccess ) {
 
 		// make sure to always kill the child process
 		try {
-			if ( child ) {
+			if ( child && !child.killed ) {
 				child.kill();
 			}
 		} catch ( e ) {}
